+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added `Activity#type_emoji` and `Activity#strava_url` - [@dblock](https://github.com/dblock).
 * Added `Athlete#name` and `Athlete#strava_url` - [@dblock](https://github.com/dblock).
 * [#5](https://github.com/dblock/strava-ruby-client/issues/5): Added support for pagination - [@dblock](https://github.com/dblock).
+* Added `Strava::Api::Client#athlete_clubs` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2018/11/23)